Reorganization of document-based classes.
